refactor/#68: PartyList 컴포넌트를 독립적으로 리팩토링
- PartyList가 내부에서 위치 정보와 파티 리스트 데이터 직접 조회
- usePartyList가 전역 스토어에서 partyType 사용하도록 변경
- MainCategory가 전역 스토어에서 partyType 사용하도록 변경
- index.tsx에서 데이터 페칭 로직 제거
